% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDbDrugIds.R
\name{getDbDrugIds}
\alias{getDbDrugIds}
\title{Get a character vector with data source-specific identifiers for
the named drug, if available.}
\usage{
getDbDrugIds(drugName, dbName)
}
\arguments{
\item{drugName}{A drug name.}

\item{dbName}{A data source chosen from "nci60", "ccle", "gdscDec15",
"ctrp", "nciSclc".}
}
\value{
A character vector with data source-specific identifiers for
the named drug, if available; otherwise returns NULL
}
\description{
Get a character vector with data source-specific identifiers for
the named drug, if available.
}
\examples{
getDbDrugIds(drugName = "Topotecan", dbName = "nci60")

}
\concept{rcellminerUtilsCDB}
