% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEmtScores.R
\name{getEmtScores}
\alias{getEmtScores}
\title{Get EMT score for a data gene expression matrix or dataframe.}
\usage{
getEmtScores(genedata, sampleOncoTree1)
}
\arguments{
\item{genedata}{A data matrix with observations along the rows and variables along the columns.}

\item{sampleOncoTree1}{a vector with tissue of origin for each sample (in genedata) based on oncoTree level 1}
}
\value{
A list with the following elements:
$emtScores pc1 scores 
$mdaAnnot metadata for the scores
$nbEmtGenes number of used EMT genes
$emtStatus final emt status
}
\description{
Get EMT score for a data gene expression matrix or dataframe.
}
\examples{
genedata <- exprs(rcellminerData::molData@eSetList$xai)
sampleOncoTree1 <- rcellminerData::molData@sampleData@samples$OncoTree1
emtResults <- getEmtScores(genedata,sampleOncoTree1)


}
\concept{
rcellminerUtilsCDB
}
