% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeneSynonyms.R
\name{getGeneSynonyms}
\alias{getGeneSynonyms}
\title{Get Gene Synonyms}
\usage{
getGeneSynonyms(gene)
}
\arguments{
\item{gene}{A gene symbol.}
}
\value{
A vector of list of synonyms including the input data
}
\description{
Get Gene Synonyms
}
\examples{
Results <- getGeneSynonyms("CGAS")


}
\concept{
rcellminerUtilsCDB
}
