% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchedCellLines.R
\name{getMatchedCellLines}
\alias{getMatchedCellLines}
\title{Get a table of matched cell lines across data sources.}
\usage{
getMatchedCellLines(dataSources)
}
\arguments{
\item{dataSources}{A character vector of (two or more) data sources chosen
from "nci60", "ccle", "cgp", "gdsc", "gdscDec15", "ctrp", "nciSclc".}
}
\value{
A data frame with matched cell lines, one column per source.
}
\description{
Get a table of matched cell lines across data sources.
}
\examples{
getMatchedCellLines(c("ccle", "cgp", "nci60"))

}
\concept{
rcellminerUtilsCDB
}
