% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPcaResults.R
\name{getPcaResults}
\alias{getPcaResults}
\title{Get principal component analysis results for a data matrix.}
\usage{
getPcaResults(X, center = TRUE, scale = TRUE)
}
\arguments{
\item{X}{A data matrix with observations along the rows and
variables along the columns.}

\item{center}{A logical value indicating whether the variables
should be shifted to be zero centered.}

\item{scale}{A logical value indicating whether the variables
should be scaled to have unit variance}
}
\value{
A list with the following elements:
$dat A data matrix with respect to the new principal component
  variables, i.e., the projections of the observed data onto
  the eigenvectors of the estimated covariance matrix.
$evecs A matrix with with variable loadings, i.e., the
  eigenvectors of the estimated covariance matrix along the
  columns.
$sdev The standard deviations of the principal components, i.e.,
  the square roots of the eigenvalues of the estimated
  covariance matrix.
$pctVar A vector indicating the percent variance associated
  with each principal component.
$cumPctVar A vector indicating the cumulative percent variance
  associated with each added principal component.
}
\description{
Get principal component analysis results for a data matrix.
}
\examples{
X <- matrix(runif(100), nrow = 25)
rownames(X) <- paste0("obs_", seq_len(nrow(X)))
colnames(X) <- paste0("var_", seq_len(ncol(X)))
pcaResults <- getPcaResults(X)

}
\concept{
rcellminerUtils
}
