% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStrDistTab.R
\name{getStrDistTab}
\alias{getStrDistTab}
\title{Returns a table of approximate matches to a set of strings, with
matches selected from a specified set.}
\usage{
getStrDistTab(seedSet, matchSet)
}
\arguments{
\item{seedSet}{A set of strings to be (approximately) matched with strings
in matchSet.}

\item{matchSet}{A set of strings that are match candidates for the
strings in seedSet.}
}
\value{
A data frame with rownames corresponding to the strings in
 seedSet, and each row specifying up to three top matches, as well
 as to the best match score.
}
\description{
Returns a table of approximate matches to a set of strings, with
matches selected from a specified set.
}
\concept{
rcellminerUtils
}
