% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueToSamplesMap.R
\name{getTissueToSamplesMap}
\alias{getTissueToSamplesMap}
\title{Get a named list mapping from tissue types to associated samples.}
\usage{
getTissueToSamplesMap(sampleData, typeLevelSeparator = ":")
}
\description{
Get a named list mapping from tissue types to associated samples.
}
\concept{rcellminerUtilsCDB}
